requirements.txt
requests
PyPDF2
docopts
